+++ include/vectutils.h
@@ -6,13 +6,13 @@
 
 namespace vu
 {
-    typedef std::vector<std::complex<float>> ComplexVect;
+    typedef std::vector<std::complex<double>> ComplexVect;
 
     extern const ComplexVect operator-(const ComplexVect &a);
     extern const ComplexVect operator+(const ComplexVect &a, const ComplexVect &b);
     extern const ComplexVect operator-(const ComplexVect &a, const ComplexVect &b);
     extern const ComplexVect operator*(const ComplexVect &a, const ComplexVect &b);
-    extern const ComplexVect operator*(const std::complex<float> &a, const ComplexVect &b);
+    extern const ComplexVect operator*(const std::complex<double> &a, const ComplexVect &b);
     extern const ComplexVect conj(const ComplexVect &a);
 
     /**
